Fixed MSVC warning
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@4815 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/common/PhysicsLX56_Projectiles.cpp
@@ -855,7 +855,7 @@ void Proj_Action::applyTo(const Proj_EventOccurInfo& eventInfo, CProjectile* prj
 		case PJ_INJUREPROJ:
 			for(Proj_EventOccurInfo::Targets::const_iterator p = eventInfo.targets.begin(); p != eventInfo.targets.end(); ++p) {
 				if(typeid(**p) == typeid(CProjectile))
-					(*p)->injure(Damage);
+					(*p)->injure((float)Damage);
 			}
 			break;
